requirements.txt
pandas>=2.0
scikit-learn>=1.0
joblib>=1.0
apache-airflow>=2.0